+++ Native/ThirdParty/MinHook/src/trampoline.c
@@ -27,6 +27,9 @@
  */
 
 #include <windows.h>
+#if defined(_MSC_VER)
+#include <intrin.h>
+#endif
 
 #ifndef ARRAYSIZE
     #define ARRAYSIZE(A) (sizeof(A)/sizeof((A)[0]))